Simplify the indexing process. We don't need to order anything.
